eslint fixes. Done and tested.
